Got the ACT_BEGIN -- ACT_END parsing idea working with the preprocessor. Explanation below.
Every time you create an actor that requires you to pass in function
parameters, you need parenthesis as well. Parenthesis are an extra token
though. So this preprocessor option lets you set function values without the
need for that extra parenthesis. This saves 1 token on most actors and parents
that are created. For a larger project, that's 50 tokes if you created 50
actors.

I would still need to incorporate the ztable this writes to in create_actor,
but that would be pretty straightforward. Also, this is based on the current
syntax of ztable. Not one that we would end up with eventually.