Changed the way filenames are sorted & added zobj improvement to template.
Instead of having a number at the beginning of filenames, the number is now at the end. Also, there can only be 1 "-" in filenames now. This makes typing files with tab completion easier. The number is needed for compilation, not for the human editor.

I also added the zobj improvement (empty key means add to table). This saved lots of compression tokens.