Small bug fixes. You can't delete an empty spot. Moves can't be duplicated now. Also, party can have empty spots anywhere now.
The way the UI works, I think it's better to allow the party to have empty spots anywhere. I allowed existing moves to show up in the move selection, that way it's impossible to delete all the moves. if you use an existing move, it is just shifted. Also, I got rid of the "switch" menu item. That's actually not too important. Just recreate your pokemon. This isn't the real pokemon game, so you're not spending dozens of hours catching and training your pkmn.