TOKEN:7083 Zobj won't create over 100 functions now (ran into lua parameter limit). Team size can be over 6 for horde.
The horde team size was 6 and updating each time a pkmn dies. But now it is 152.

Whirlwind/roar/teleport will work across the entire team, so you can use those to unlock missingno (if you can't actually beat the whole horde).

Also made the "live" count when you finish a fight/view team stats ignore missingno, so there is more mystery to it.