TOKEN:8116 Game can be minified now. Minifier has stable compilation. League teams are now stored in cartridge.
With the minifier & storing the league teams in the cartridge, I saved around 4000 compression tokens.

I will need to store moves in the cartridge too as I'm still over on compression space.

That will likely write over the pkmn cries.

When sprites are done, we may need to limit cries from 4 to 2 notes.